Restructure tests and assert values
